inputs in common
