+++ features/ajna/borrow/controls/AjnaBorrowFormController.tsx
@@ -3,8 +3,8 @@ import { AjnaBorrowFormContentDeposit } from 'features/ajna/borrow/sidebars/Ajna
 import { AjnaBorrowFormContentManage } from 'features/ajna/borrow/sidebars/AjnaBorrowFormContentManage'
 import { useAjnaGeneralContext } from 'features/ajna/common/contexts/AjnaGeneralContext'
 import { useAjnaProductContext } from 'features/ajna/common/contexts/AjnaProductContext'
-import { AjnaFormContentRisk } from 'features/ajna/common/sidebars/AjnaBorrowFormContentRisk'
-import { AjnaFormContentTransaction } from 'features/ajna/common/sidebars/AjnaBorrowFormContentTransaction'
+import { AjnaFormContentRisk } from 'features/ajna/common/sidebars/AjnaFormContentRisk'
+import { AjnaFormContentTransaction } from 'features/ajna/common/sidebars/AjnaFormContentTransaction'
 import { AjnaFormView } from 'features/ajna/common/views/AjnaFormView'
 import { useTranslation } from 'next-i18next'
 import React from 'react'
