+++ features/ajna/borrow/sidebars/AjnaBorrowFormContentGenerate.tsx
@@ -1,23 +1,25 @@
 import { AjnaBorrowFormContentSummary } from 'features/ajna/borrow/sidebars/AjnaBorrowFormContentSummary'
-import { AjnaBorrowFormFieldGenerate } from 'features/ajna/borrow/sidebars/AjnaBorrowFormFields'
-import { AjnaFormFieldDeposit } from 'features/ajna/common/components/AjnaFormFieldDeposit'
 import { useAjnaProductContext } from 'features/ajna/common/contexts/AjnaProductContext'
+import {
+  AjnaFormFieldDeposit,
+  AjnaFormFieldGenerate,
+} from 'features/ajna/common/sidebars/AjnaFormFields'
 import React from 'react'
 
 export function AjnaBorrowFormContentGenerate() {
   const {
     form: {
-      state: { depositAmount, depositAmountUSD, generateAmount },
+      dispatch,
+      state: { generateAmount },
     },
   } = useAjnaProductContext('borrow')
 
   return (
     <>
-      <AjnaBorrowFormFieldGenerate resetOnClear />
+      <AjnaFormFieldGenerate dispatchAmount={dispatch} resetOnClear />
       <AjnaFormFieldDeposit
+        dispatchAmount={dispatch}
         isDisabled={!generateAmount || generateAmount?.lte(0)}
-        depositAmount={depositAmount}
-        depositAmountUSD={depositAmountUSD}
       />
       {generateAmount && <AjnaBorrowFormContentSummary />}
     </>
