+++ features/ajna/borrow/sidebars/AjnaBorrowFormContentPayback.tsx
@@ -1,22 +1,26 @@
 import { AjnaBorrowFormContentSummary } from 'features/ajna/borrow/sidebars/AjnaBorrowFormContentSummary'
-import {
-  AjnaBorrowFormFieldPayback,
-  AjnaBorrowFormFieldWithdraw,
-} from 'features/ajna/borrow/sidebars/AjnaBorrowFormFields'
 import { useAjnaProductContext } from 'features/ajna/common/contexts/AjnaProductContext'
+import {
+  AjnaFormFieldPayback,
+  AjnaFormFieldWithdraw,
+} from 'features/ajna/common/sidebars/AjnaFormFields'
 import React from 'react'
 
 export function AjnaBorrowFormContentPayback() {
   const {
     form: {
+      dispatch,
       state: { paybackAmount },
     },
   } = useAjnaProductContext('borrow')
 
   return (
     <>
-      <AjnaBorrowFormFieldPayback resetOnClear />
-      <AjnaBorrowFormFieldWithdraw isDisabled={!paybackAmount || paybackAmount?.lte(0)} />
+      <AjnaFormFieldPayback dispatchAmount={dispatch} resetOnClear />
+      <AjnaFormFieldWithdraw
+        dispatchAmount={dispatch}
+        isDisabled={!paybackAmount || paybackAmount?.lte(0)}
+      />
       {paybackAmount && <AjnaBorrowFormContentSummary />}
     </>
   )
