+++ features/ajna/earn/controls/AjnaEarnFormController.tsx
@@ -1,8 +1,8 @@
 import { getToken } from 'blockchain/tokensMetadata'
 import { useAjnaGeneralContext } from 'features/ajna/common/contexts/AjnaGeneralContext'
 import { useAjnaProductContext } from 'features/ajna/common/contexts/AjnaProductContext'
-import { AjnaFormContentRisk } from 'features/ajna/common/sidebars/AjnaBorrowFormContentRisk'
-import { AjnaFormContentTransaction } from 'features/ajna/common/sidebars/AjnaBorrowFormContentTransaction'
+import { AjnaFormContentRisk } from 'features/ajna/common/sidebars/AjnaFormContentRisk'
+import { AjnaFormContentTransaction } from 'features/ajna/common/sidebars/AjnaFormContentTransaction'
 import { AjnaFormView } from 'features/ajna/common/views/AjnaFormView'
 import { AjnaEarnFormContentDeposit } from 'features/ajna/earn/sidebars/AjnaEarnFormContentDeposit'
 import { useTranslation } from 'next-i18next'
